requirements.txt
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asciinema==2.0.2
asciitree==0.3.3
asgiref==3.5.2
attrs==19.3.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
boto==2.49.0
Brlapi==0.7.0
cached-property==1.5.1
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
crcmod==1.7
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dblatex===0.3.11py3
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.4
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
evdev==1.3.0
fasteners==0.14.1
filelock==3.0.12
fonttools==4.29.1
future==0.18.2
galternatives==1.0.6
gunicorn==20.1.0
gurux-common==1.0.13
gurux-dlms==1.0.107
gurux-net==1.0.17
gurux-serial==1.0.15
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
joblib==1.1.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libiio==0.19
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.5.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.21.2
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.8
packaging==21.3
pandas==1.4.0
paramiko==2.6.0
pdfrw==0.4
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.7
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-magic==0.4.16
pytz==2021.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
scikit-learn==1.1.1
scipy==1.8.0
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.8.0
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
smeterd==2.9.0
soupsieve==1.9.5
sqlparse==0.4.2
ssh-import-id==5.10
systemd-python==234
texttable==1.6.2
threadpoolctl==3.1.0
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uploads==0.8
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
webencodings==0.5.1
websocket-client==0.53.0
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
zipp==1.0.0
